backend/requirements.txt
# ============================================================
# FastAPI & Web Framework
# ============================================================
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
aiofiles

# ============================================================
# Database & ORM
# ============================================================
sqlalchemy
psycopg2-binary
alembic

# ============================================================
# Authentication & Security
# ============================================================
pydantic
pydantic[email]
python-jose[cryptography]
pyjwt

# ============================================================
# Audio Processing & AI/ML
# ============================================================
torch
torchaudio
transformers
openai-whisper
whisperx
pyannote.audio
sounddevice
soundfile
numpy
librosa

# ============================================================
# Zoom SDK & API Integration
# ============================================================
requests
aiohttp
selenium
webdriver-manager

# ============================================================
# Utilities & Configuration
# ============================================================
python-dotenv
fastapi-mail
selenium
webdriver-manager

# ============================================================
# Testing
# ============================================================
pytest
pytest-mock
pytest-cov
pytest-asyncio

# ============================================================
# Development Tools
# ============================================================
black
flake8
mypy
